+++ front_end/lib/ui/ramp_up_home_page/ramp_up_home_page_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:front_end/ui/ramp_up_home_page/ramp_up_home_page_bloc.dart';
-import 'package:front_end/ui/ramp_up_home_page/ramp_up_page_view.dart';
+import 'package:front_end/ui/ramp_up_home_page/ramp_up_home_page_view.dart';
 
 class RampUpHomeScreenProvider extends BlocProvider<RampUpHomeScreenBloc> {
   RampUpHomeScreenProvider({
