+++ front_end/lib/ui/ramp_up_home_page/ramp_up_home_page_view.dart
